+++ src/components/internal/index.ts
@@ -3,3 +3,5 @@ export { default as HelpAndErrorTexts } from './HelpAndErrorTexts.vue';
 export { BlockWrapper, type TBlockWrapperProps } from './BlockWrapper';
 export { ModalBackdropWrapper } from './ModalBackdropWrapper';
 export { default as Loader } from './Loader.vue';
+export { default as LinkButton } from './LinkButton.vue';
+export { default as Mesh } from './Mesh.vue';